Move XML -> JSON into parse.rs
